Move CoreTest to CoreWPFTest
